move/cleanup message console, don't understand purpose of Message Console class
